applying viewer mode
